+++ commet/lib/ui/molecules/timeline_events/events/timeline_event_view_message.dart
@@ -40,9 +40,9 @@ class TimelineEventViewMessage extends StatefulWidget {
       this.jumpToEvent,
       this.readReceipts = const [],
       this.onReadReceiptsTapped,
+      this.onDoubleTapMessage,
       this.detailed = false,
       this.previewMedia = false,
-      this.setReplyingEvent,
       required this.initialIndex});
 
   final Function(String eventId)? jumpToEvent;
@@ -57,7 +57,7 @@ class TimelineEventViewMessage extends StatefulWidget {
   final bool isThreadTimeline;
   final bool previewMedia;
   final Function()? onReadReceiptsTapped;
-  final Function(TimelineEvent? event)? setReplyingEvent;
+  final Function()? onDoubleTapMessage;
 
   @override
   State<TimelineEventViewMessage> createState() =>
@@ -70,7 +70,8 @@ class _TimelineEventViewMessageState extends State<TimelineEventViewMessage>
   late String senderId;
   late Color senderColor;
 
-  late TimelineEvent<Client> event;
+  late bool mentionsRoom;
+  late List<String> mentions;
 
   String get messageFailedToDecrypt => Intl.message("Failed to decrypt event",
       desc: "Placeholde text for when a message fails to decrypt",
@@ -135,9 +136,9 @@ class _TimelineEventViewMessageState extends State<TimelineEventViewMessage>
       formattedContent: formattedContent,
       timestamp: timestampToString(sentTime),
       edited: edited,
-      isMentioningSelf: event.mentionsRoom ||
-          event.mentions.contains(widget.timeline!.client.self!.identifier),
-      onDoubleTapMessage: () => widget.setReplyingEvent?.call(event),
+      isMentioningSelf: mentionsRoom ||
+          mentions.contains(widget.timeline!.client.self!.identifier),
+      onDoubleTapMessage: widget.onDoubleTapMessage,
       avatarBuilder: (child) {
         var room = widget.room ?? widget.timeline?.room;
 
@@ -227,7 +228,8 @@ class _TimelineEventViewMessageState extends State<TimelineEventViewMessage>
   }
 
   void loadStateFromEvent(TimelineEvent event) {
-    this.event = event;
+    mentionsRoom = event.mentionsRoom;
+    mentions = event.mentions;
     showSender = shouldShowSender(index);
     var room = widget.room ?? widget.timeline?.room;
