+++ commet/lib/ui/molecules/timeline_events/timeline_view_entry.dart
@@ -24,7 +24,6 @@ import 'package:commet/ui/molecules/timeline_events/timeline_event_menu_dialog.d
 import 'package:commet/ui/molecules/user_panel.dart';
 import 'package:commet/ui/navigation/adaptive_dialog.dart';
 import 'package:flutter/material.dart';
-import 'package:intl/intl.dart';
 import 'package:tiamat/atoms/context_menu.dart';
 import 'package:tiamat/tiamat.dart' as tiamat;
 
@@ -41,6 +40,7 @@ class TimelineViewEntry extends StatefulWidget {
       this.isThreadTimeline = false,
       this.previewMedia = false,
       this.highlightedEventId,
+      this.overrideShowSender,
       super.key});
   final Timeline timeline;
   final int initialIndex;
@@ -52,6 +52,7 @@ class TimelineViewEntry extends StatefulWidget {
   final bool isThreadTimeline;
   final String? highlightedEventId;
   final bool previewMedia;
+  final bool? overrideShowSender;
 
   // Should be true if we are showing this event on its own, and not as part of a timeline
   final bool singleEvent;
@@ -93,16 +94,11 @@ class TimelineViewEntryState extends State<TimelineViewEntry>
 
   late DateTime time;
   bool showDateSeperator = false;
-  bool showUnreadMarker = false;
 
   ThreadsComponent? threads;
   PollComponent? polls;
 
   List<String> readReceipts = [];
-
-  String get labelTimelineNewMessagesMarker => Intl.message("New messages",
-      desc: "Text that is shown below the last read message",
-      name: "labelTimelineNewMessagesMarker");
 
   @override
   void initState() {
@@ -136,7 +132,6 @@ class TimelineViewEntryState extends State<TimelineViewEntry>
     _widgetType = eventToDisplayType(event, polls: polls);
 
     showDateSeperator = shouldEventShowDate(eventIndex);
-    showUnreadMarker = shouldEventShowUnreadMarker(eventIndex);
     highlighted = event.eventId == widget.highlightedEventId;
   }
 
@@ -200,25 +195,6 @@ class TimelineViewEntryState extends State<TimelineViewEntry>
         2) return true;
 
     return false;
-  }
-
-  bool shouldEventShowUnreadMarker(int index) {
-    final events = widget.timeline.events;
-
-    bool isHidden(event) =>
-        eventToDisplayType(event, polls: polls) ==
-        TimelineEventWidgetDisplayType.hidden;
-
-    if (index == 0 || events.take(index).every(isHidden)) return false;
-
-    final lastReadIndex =
-        events.indexWhere((e) => e.eventId == widget.timeline.room.lastRead);
-
-    if (lastReadIndex > index) return false;
-    if (lastReadIndex == index) return true;
-    if (!isHidden(events[lastReadIndex])) return false;
-
-    return events.getRange(lastReadIndex, index).every(isHidden);
   }
 
   @override
@@ -419,34 +395,6 @@ class TimelineViewEntryState extends State<TimelineViewEntry>
       );
     }
 
-    if (showUnreadMarker)
-      result = Column(children: [
-        result,
-        Row(
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            Expanded(
-                child: Divider(
-                    color: Colors.red,
-                    thickness: 4.0,
-                    radius: BorderRadiusGeometry.horizontal(
-                        right: Radius.circular(16.0)))),
-            Padding(
-              padding: EdgeInsets.symmetric(horizontal: 15),
-              child: Text(labelTimelineNewMessagesMarker,
-                  style: TextStyle(
-                      color: Colors.red, fontWeight: FontWeight.bold)),
-            ),
-            Expanded(
-                child: Divider(
-                    color: Colors.red,
-                    thickness: 4.0,
-                    radius: BorderRadiusGeometry.horizontal(
-                        left: Radius.circular(16.0)))),
-          ],
-        ),
-      ]);
-
     return result;
   }
 
@@ -461,6 +409,7 @@ class TimelineViewEntryState extends State<TimelineViewEntry>
       return null;
     }
 
+    var event = widget.timeline.tryGetEvent(eventId);
     if (_widgetType == TimelineEventWidgetDisplayType.message)
       return TimelineEventViewMessage(
           key: eventKey,
@@ -468,8 +417,10 @@ class TimelineViewEntryState extends State<TimelineViewEntry>
           isThreadTimeline: widget.isThreadTimeline,
           detailed: widget.showDetailed || selected,
           onReadReceiptsTapped: onReadReceiptsTapped,
+          onDoubleTapMessage: () => widget.setReplyingEvent?.call(event),
           readReceipts: readReceipts,
-          overrideShowSender: widget.singleEvent || showDateSeperator,
+          overrideShowSender: widget.overrideShowSender ??
+              (widget.singleEvent || showDateSeperator),
           jumpToEvent: widget.jumpToEvent,
           previewMedia: widget.previewMedia,
           initialIndex: widget.initialIndex);
